requirements.txt
django
djangorestframework
django-phonenumber-field[phonenumbers]
djangorestframework-simplejwt
drf-yasg
django-cors-headers
psycopg2-binary
Pillow
django-mail-templated